Split API into abstract base classes
